dialog box added to the add and remove from login flow
